Refactor file names and makefile
